requirements.txt
torch
torchvision
numpy
opencv-python
scikit-learn
matplotlib
pillow
tqdm
pyyaml
pydensecrf
